.kb/pyproject.toml
[project]
name = "kb-mcp"
version = "0.1.0"
description = "MCP server for personal knowledge base — search, explore, and navigate 200+ atomic notes"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.27.0",
    "numpy>=2.4.4",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
]
